requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.0
billiard==3.5.0.5
bs4==0.0.1
cachetools==3.1.1
celery==4.1.1
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
elasticsearch==6.4.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
idna==2.8
kombu==4.6.3
lxml==4.4.0
oauth2client==4.1.3
parsel==1.5.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2019.2
redis==3.3.5
requestium==0.1.9
requests==2.22.0
requests-file==1.4.3
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
tldextract==2.2.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
w3lib==1.20.0
